refactor: removed SectionWrapper component
